requirements.txt
asgiref==3.3.4
autopep8==1.5.7
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.2.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
qrcode==6.1
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
Unipath==1.1
whitenoise==5.2.0
